Fix financial-year accuracy and harden the Go pipeline
Verifying the Go port against the original Python revealed several
financially material discrepancies. Fixed test-first, plus robustness
and testing improvements:

Correctness:
- ParseFinancialYear returned Jan 1 / Jan 1 instead of Apr 1 / Mar 31,
  corrupting Schedule CG (dropped Jan-Mar sales, included prior-FY sales)
  and Schedule AL (dropped grants vested Jan-Mar).
- Schedule FA now uses the calendar year (1 Jan - 31 Dec of the FY start)
  for filtering, the peak-value window, and the year-end closing price and
  forex rate. A PriceProvider interface makes it testable.
- Share quantities parsed as float64 (were truncated to int).
- Award/grant numbers kept as strings (identifiers, not integers).

Robustness:
- File parsing fails loud on malformed cells; date errors name the column
  and expected format; duplicate required columns are rejected.
- SBI forex lookback capped at a named 15-day constant with an explicit
  error when exceeded.
- Yahoo prices decoded as *float64 so null days are skipped (not treated
  as 0.0); non-200 responses fail loud while still surfacing the API error.

Testing:
- Yahoo client tests converted to httptest (offline, deterministic).
- Added exporter round-trip test and boundary tests for CG/AL/FA.
- Added an opt-in end-to-end test (RUN_E2E=1) that runs the full pipeline
  against anonymized ETrade fixtures and compares the generated workbook
  cell-by-cell against a committed golden workbook.
- README documents the test suite and fixtures.

Tooling:
- Bump Go to 1.26.0 (go.mod and .tool-versions).
- Update direct dependencies: excelize v2.8.0 -> v2.11.0,
  testify v1.8.0 -> v1.11.1.
- CI installs Go from go.mod (go-version-file) so the toolchain matches
  and coverage does not fail with "no such tool covdata"; pin
  actions/setup-go to v6.5.0 and actions/checkout to v7.

+++ internal/schedule/assets_and_liabilities.go
@@ -44,7 +44,7 @@ func GenerateScheduleAL(
 
 		record := AssetsAndLiabilitiesRecord{
 			ShareRecord:       share,
-			CostOfAcquisition: float64(share.SharesIssued) * share.FMVPerShare * issueRate.TTBuyExchangeRate,
+			CostOfAcquisition: share.SharesIssued * share.FMVPerShare * issueRate.TTBuyExchangeRate,
 			CostOfImprovement: 0,
 		}
